requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
xlsxwriter>=3.1.0
numpy>=1.24.0
anthropic>=0.40.0
python-dotenv>=1.0.0
yfinance>=0.2.0
akshare>=1.10.0
requests>=2.31.0
